Refactor app types in internal files, remove mongoose
