+++ package.json
@@ -11,14 +11,12 @@
   "dependencies": {
     "@prisma/client": "^4.0.0",
     "axios": "^0.27.2",
-    "mongoose": "^6.4.4",
     "next": "^12.2.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
-    "@types/mongoose": "^5.11.97",
     "@types/node": "18.0.3",
     "@types/react": "18.0.15",
     "@types/react-dom": "18.0.6",
